requirements.txt
pytesseract==0.3.10
Pillow==10.0.1
pyzxing==0.1
tqdm==4.66.1